arena

+++ src/just.c
@@ -1,10 +1,11 @@
 #include "just.h"
 #include <stdarg.h>
+#include <stdint.h>
 
 // arena -------------------------------------------------------------------------------------------
 arena*
-arena_create(u64 capacity) {
-        assert(sizeof(arena) <= UINT64_MAX - capacity);
+arena_create(size_t capacity) {
+        assert(sizeof(arena) <= SIZE_MAX - capacity);
         arena* a = malloc(sizeof(arena) + capacity);
         assert(a != NULL);
         *a = (arena){.cap = capacity};
@@ -12,11 +13,11 @@ arena_create(u64 capacity) {
 }
 
 void*
-arena_alloc(arena* a, u64 align, u64 size, u64 count) {
+arena_alloc(arena* a, size_t align, size_t size, size_t count) {
         assert(a != NULL && "Invalid arena");
         assert(ISPOW2(align) && "Invalid alignment");
         assert(size != 0 && "Invalid size");
-        u64 padding = -(uintptr_t)(a->data + a->used) & (align - 1);
+        size_t padding = -(uintptr_t)(a->data + a->used) & (align - 1);
         assert(count <= (a->cap - a->used - padding) / size && "Increase arena capacity");
         void* ptr = a->data + a->used + padding;
         a->used += padding + count * size;
